Added property to correct mirrored touchscreen
A property was created to fix touchscreen problems when the screen is
mirrored. That is, a touch in certain position is identified by a touch
in a mirrored position.
The user can modify this property if it is identified that the
touchscreen is mirrored relative to x, y, or both.
Adapted the Simple Views example to support this change.